More fixes to prevent concurrent m_owner writing
Setting m_owner in ::wait() is not useful and wastes cpu cycles/creates opportunity for both cores to write to it at the same time, better to just remove it. 

We should also set m_owner before giving the semaphore in ::give() to be sure it's not accessed by both cores after the semaphore is given.